Put in various portability changes to compile on Solaris w/o warnings. The symptoms came courtesy of David Lee.
CVS patchset: 572
CVS date: 2001/06/07 21:29:44

+++ heartbeat/api_test.c
@@ -81,7 +81,7 @@ main(int argc, char ** argv)
 	(void)_ha_msg_h_Id;
 
 	hb = ll_cluster_new("heartbeat");
-	fprintf(stderr, "PID=%d\n", getpid());
+	fprintf(stderr, "PID=%ld\n", (long)getpid());
 	fprintf(stderr, "Signing in with heartbeat\n");
 	if (hb->llc_ops->signon(hb, "ping")!= HA_OK) {
 		fprintf(stderr, "Cannot sign on with heartbeat\n");
